+++ create_request_info.py
@@ -1,7 +1,6 @@
 import db_conn
 import os
 import json
-import pickle
 import utility
 import customized_settings